requirements-dev.txt
# Development-only dependencies
ruff==0.6.7
pytest-django==4.8.0
